requirements.txt
### dependency
tqdm
dvc
pandas
numpy
mlflow
argon2-cffi
argon2-cffi-bindings
attrs
backcall
bleach
certifi
cffi
click
colorama
cycler
debugpy
decorator
defusedxml
entrypoints
fonttools
ghp-import
importlib-metadata
importlib-resources
ipykernel
ipython
ipython-genutils
jedi
Jinja2
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
kiwisolver
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mergedeep
mistune
mkdocs
mkdocs-material
mkdocs-material-extensions
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
packaging
pandas
pandocfilters
parso
pickleshare
Pillow
prometheus-client
prompt-toolkit
pycparser
Pygments
pymdown-extensions
pyparsing
pyrsistent
python-dateutil
pytz
pywin32
pywinpty
PyYAML
pyyaml_env_tag
pyzmq
scipy
seaborn
Send2Trash
six
sklearn
scikit-learn
terminado
testpath
torch
torchaudio
torchvision
tornado
traitlets
typing_extensions
watchdog
wcwidth
webencodings
wincertstore
zipp
prettytable

 -e . # << uncomment this to install from setup.py
